verificacion de expiracion de sesion cada hora, autorizacion de seguridad para request listos

+++ src/migrations/20240210193230-create-user.js
@@ -19,7 +19,7 @@ module.exports = {
         type: Sequelize.DataTypes.INTEGER
       },
       numberId: {
-        type: Sequelize.DataTypes.INTEGER
+        type: Sequelize.DataTypes.BIGINT
       },
       dateOfBirth: { 
         allowNull: true,
